Some adjustments and bugfix

+++ src/bitcode/parsing_context.cpp
@@ -12,19 +12,19 @@ namespace bitcode {
         return !block_infos.empty();
     }
 
-    BlockInfo* ParsingContext::GetBlockInfo(int block_id) {
+    BlockInfo* ParsingContext::GetBlockInfo(uint32_t block_id) {
         if (!block_infos.empty() && block_infos.back().block_id == block_id)
             return &block_infos.back();
 
-        int size = (int)block_infos.size();
-        for (int i = 0; i < size; i++) {
+        size_t size = block_infos.size();
+        for (size_t i = 0; i < size; i++) {
             if (block_infos[i].block_id == block_id)
                 return &block_infos[i];
         }
         return nullptr;
     }
 
-    BlockInfo* ParsingContext::GetOrCreateBlockInfo(int block_id) {
+    BlockInfo* ParsingContext::GetOrCreateBlockInfo(uint32_t block_id) {
         BlockInfo* info = GetBlockInfo(block_id);
         if (info)
             return info;
